renaming located_object_handlers as common_ui_targets
