+++ traitsui/testing/tester/qt4/common_ui_targets.py
